Extract bounty attempt routes
